\name{print.snn}
\alias{print.snn}

\title{
S3 Method for Printing an 'snn' Object
}
\description{
Print the survival neural network.
}

\usage{
\method{print}{snn}(x,  ...)
}

\arguments{
  \item{x}{An object returned by the function \code{SNN}.}
  \item{...}{Further arguments passed.}
}

\seealso{
  \code{\link{SNN}}.
}

\value{No return value for this S3 method.}

\keyword{Printing}
